refactor(auth): extract keepalive + refresh into _auth subpackage with facade-forwarding (tier-10 PR-B-high) — HIGH-RISK
Shrink ``src/notebooklm/auth.py`` from 1564 LOC to 722 LOC by extracting the
keepalive ``RotateCookies`` poke and the refresh-cmd / token-fetch flow into
two new private modules. The compatibility surface on ``notebooklm.auth``
stays byte-identical for first-party callers (CLI, ``_core.py``,
``_core_lifecycle.py``, tests) AND for white-box monkeypatch sites.

## New modules
- ``src/notebooklm/_auth/keepalive.py`` (355 LOC): owns
  ``KEEPALIVE_ROTATE_URL``, the ``_POKE_LOCKS_BY_LOOP`` / ``_LAST_POKE_ATTEMPT_MONOTONIC``
  state, ``_get_poke_lock`` / ``_try_claim_rotation`` / ``_file_lock_try_exclusive``,
  ``_is_recently_rotated``, and ``_poke_session`` / ``_rotate_cookies``.
- ``src/notebooklm/_auth/refresh.py`` (737 LOC): owns the
  ``_REFRESH_LOCKS_BY_LOOP`` / ``_REFRESH_INFLIGHT_BY_LOOP`` registries,
  ``_coalesced_run_refresh_cmd``, ``_run_refresh_cmd`` (carrying the tier-9 E
  P1-18 two-channel disclosure redaction verbatim), ``_fetch_tokens_with_refresh``,
  ``_fetch_tokens_with_jar``, and the public ``fetch_tokens`` /
  ``fetch_tokens_with_domains`` entry points.
- ``src/notebooklm/_auth/cookies.py`` gains ``_update_cookie_input`` (option a
  per plan — cohesive with ``flatten_cookie_map`` which it consumes).

## Facade write-through (``_AuthFacadeModule``)
Two new name-sets route reads + mirror writes to the moved modules:
- ``_AUTH_KEEPALIVE_FACADE_NAMES`` (15 names) → ``_auth.keepalive``
- ``_AUTH_REFRESH_FACADE_NAMES`` (17 names) → ``_auth.refresh``

Plus two cross-module dependency mirror sets for names that physically live
in one seam but are aliased into another as rebindable bare-name slots:
- ``_REFRESH_DEP_MIRROR_NAMES``: ``snapshot_cookie_jar``,
  ``save_cookies_to_storage``, ``build_httpx_cookies_from_storage``,
  ``build_cookie_jar``, ``_replace_cookie_jar``, ``_cookie_map_from_jar``,
  ``flatten_cookie_map``, ``_update_cookie_input``, ``extract_csrf_from_html``,
  ``extract_session_id_from_html``, ``_safe_url``,
  ``_resolve_token_route_kwargs``, ``get_storage_path``.
- ``_KEEPALIVE_DEP_MIRROR_NAMES``: ``_file_lock``, ``_rotation_lock_path``.

The ``_poke_session`` slot is also mirrored from auth → ``_auth.refresh``
explicitly so ``_fetch_tokens_with_jar`` (now in refresh.py) observes any
monkeypatch placed on ``notebooklm.auth._poke_session``.

## Invariants preserved
- ``_LAST_POKE_ATTEMPT_MONOTONIC`` and ``_POKE_LOCKS_BY_LOOP`` are shared by
  IDENTITY (direct-assignment) between ``notebooklm.auth`` and
  ``_auth.keepalive`` so ``tests/conftest.py`` ``.clear()`` mutations through
  the facade propagate into the moved bodies.
- ``auth.subprocess`` re-export preserved (white-box tests patch
  ``auth_mod.subprocess.run`` per ``test_refresh_cmd_shlex.py:73``).
- Tier-9 E (P1-18) ``_run_refresh_cmd`` redaction carried verbatim: the
  user-facing ``RuntimeError`` surfaces only exit code + executable basename;
  full stdout/stderr only reach the DEBUG logger.
- ``_settle`` callback in ``_coalesced_run_refresh_cmd`` intentionally leaves
  the done future in the registry (CodeRabbit PR #621 finding) so the
  ``except CancelledError`` branch can still inspect ``inflight.exception()``.

## Tests
- ``tests/unit/test_public_shims.py``: 12 new SEMANTIC patch-and-execute
  tests that patch via ``auth_mod.X = fake`` and assert the patch rebinds the
  call site inside the moved module. Covers all five highest-risk monkeypatch
  targets from the signoff list: ``_run_refresh_cmd``, ``_fetch_tokens_with_jar``,
  ``_get_refresh_lock``, ``_poke_session``, ``_rotate_cookies``. Plus
  cross-module-mirror coverage for ``snapshot_cookie_jar`` and
  ``build_httpx_cookies_from_storage`` and facade-name-set completeness
  identity pins.

## Verification
- 5370 unit + integration tests pass (5358 baseline + 12 new semantic tests).
- mypy clean.
- pre-commit (ruff lint + format) clean.
- All 8 signoff items verified (a-h).
