+++ src/notebooklm/_auth/__init__.py
@@ -1,5 +1,5 @@
 """Private authentication implementation package."""
 
-from . import extraction, headers, paths
+from . import extraction, headers, keepalive, paths, refresh
 
-__all__ = ["extraction", "headers", "paths"]
+__all__ = ["extraction", "headers", "keepalive", "paths", "refresh"]
